v4.0.2
Replaced BRIGHT and FLASH commands with CLUT and COLOR.
Replaced LPRINT and LLIST commands with SLOW and FAST.
Replaced the old COLOR command with the MODE command.
Brought back the CALL command with its original token.
Renamed the TYPE command to UDG, as per the TK90X.
BORDER no longer sets the hi-res color combo.
In hi-res mode PEN and PAPER have a maximum value of 7.